requirements.txt
pyyaml>=6.0
tokenizers>=0.20
torch>=2.4
transformers>=4.52
tqdm>=4.66
safetensors>=0.4
anthropic>=0.25
openai>=1.30
groq>=0.5
sentence-transformers>=3.0
apscheduler>=3.10
faiss-cpu>=1.8
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
gguf>=0.10
